Log sql error on db lock, dont quit

+++ src/p_sqldump.c
@@ -31,35 +31,38 @@
 #include <p_sqldump.h>
 #include <p_tools.h>
 #include <stdlib.h>
+#include <log.h>
 
-static void exitOnError(sqlite3 * db, const char * where, int sqerr)
+#define LOGFD  peer[id].logfd 
+
+static void exitOnError(FILE * logfd, sqlite3 * db, const char * where, int sqerr)
 {
     if (sqerr!=SQLITE_OK && sqerr != SQLITE_ROW && sqerr != SQLITE_DONE )
 	{
-		fprintf(stderr,"ERROR!! at %s\n", where);
-		fprintf(stderr,"      > sql code       %d\n",sqerr);
-		fprintf(stderr,"      > sql error      %d\n",sqlite3_errcode(db));
-		fprintf(stderr,"      > sql exterror   %d\n",sqlite3_extended_errcode(db));
-		fprintf(stderr,"      > sql msg        %s\n",sqlite3_errmsg(db));
-		fprintf(stderr,"      > sql errstring  %s\n",sqlite3_errstr(sqerr));
-		fprintf(stderr,"Sorry, have to quit (-2), bye\n");
+		log_error(logfd,"ERROR!! at %s", where);
+		log_error(logfd,"      > sql code       %d",sqerr);
+		log_error(logfd,"      > sql error      %d",sqlite3_errcode(db));
+		log_error(logfd,"      > sql exterror   %d",sqlite3_extended_errcode(db));
+		log_error(logfd,"      > sql msg        %s",sqlite3_errmsg(db));
+		log_error(logfd,"      > sql errstring  %s",sqlite3_errstr(sqerr));
+		log_error(logfd,"Sorry, have to quit (-2), bye");
 		sqlite3_close(db);
 		exit(-2);
 	}
 }
 
-static int  returnOnError(sqlite3 * db, const char * where, int sqerr, sqlite3_stmt * pstmt)
+static int  returnOnError(FILE * logfd, sqlite3 * db, const char * where, int sqerr, sqlite3_stmt * pstmt)
 {
     if (sqerr!=SQLITE_OK && sqerr != SQLITE_ROW && sqerr != SQLITE_DONE )
 	{
-		fprintf(stderr,"WARNING!! at %s\n", where);
-		fprintf(stderr,"      > sql code       %d\n",sqerr);
-		fprintf(stderr,"      > sql error      %d\n",sqlite3_errcode(db));
-		fprintf(stderr,"      > sql exterror   %d\n",sqlite3_extended_errcode(db));
-		fprintf(stderr,"      > sql msg        %s\n",sqlite3_errmsg(db));
-		fprintf(stderr,"      > sql errstring  %s\n",sqlite3_errstr(sqerr));
-		sqlite3_finalize(pstmt);
-		sqlite3_close(db);
+		log_error(logfd,"WARNING!! at %s", where);
+		log_error(logfd,"      > sql code       %d",sqerr);
+		log_error(logfd,"      > sql error      %d",sqlite3_errcode(db));
+		log_error(logfd,"      > sql exterror   %d",sqlite3_extended_errcode(db));
+		log_error(logfd,"      > sql msg        %s",sqlite3_errmsg(db));
+		log_error(logfd,"      > sql errstring  %s",sqlite3_errstr(sqerr));
+		if (pstmt) 
+			sqlite3_finalize(pstmt);
 		return 1;
 	} else {
 		return 0;
@@ -72,6 +75,7 @@ void p_sqldump_open_file(struct peer_t *peer, int id, struct timeval *ts)
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	snprintf(peer[id].sqldbname, sizeof(peer[id].sqldbname), "%s/%s_routes.db.sqlite3",
 		DUMPDIR,
@@ -88,8 +92,19 @@ void p_sqldump_open_file(struct peer_t *peer, int id, struct timeval *ts)
 	}
 
 
+	/* create log file upfront for this peer, we need this to run in prodn */
+	char logfile[256];
+	snprintf(logfile, sizeof(logfile),"%s/log.%s",
+		PEERLOGDIR,
+		peer[id].af == 4 ? p_tools_ip4str(id, &peer[id].ip4) : p_tools_ip6str(id, &peer[id].ip6));
+
+	peer[id].logfd = fopen(logfile, "wb" );
+	log_info( LOGFD ,  "Opened log file for peer %s", 
+		peer[id].af == 4 ? p_tools_ip4str(id, &peer[id].ip4) : p_tools_ip6str(id, &peer[id].ip6));
+
+
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE|SQLITE_OPEN_CREATE, NULL);
-	exitOnError(pSQL3, "open", sqerr);
+	exitOnError(LOGFD,pSQL3, "open", sqerr);
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
@@ -101,7 +116,7 @@ void p_sqldump_open_file(struct peer_t *peer, int id, struct timeval *ts)
 						"    NEXTHOP					VARCHAR,    "
                         "    TIMESTAMP	 				INTEGER     "
 						"    );", NULL,NULL,NULL);
-	exitOnError(pSQL3, "create table prefix_paths_v4", sqerr);
+	exitOnError(LOGFD,pSQL3, "create table prefix_paths_v4", sqerr);
 
 	// prefix-v6 -> as path, the current view 
 	sqerr=sqlite3_exec(pSQL3, "CREATE TABLE IF NOT EXISTS PREFIX_PATHS_V6 (    "
@@ -111,35 +126,34 @@ void p_sqldump_open_file(struct peer_t *peer, int id, struct timeval *ts)
 						"    NEXTHOP					VARCHAR,    "
                         "    TIMESTAMP	 				INTEGER     "
 						"    );", NULL,NULL,NULL);
-	exitOnError(pSQL3, "create table prefix_paths_v6", sqerr);
+	exitOnError(LOGFD,pSQL3, "create table prefix_paths_v6", sqerr);
 
 	// last keep alive 
 	sqerr=sqlite3_exec(pSQL3, "CREATE TABLE IF NOT EXISTS LAST_KEEPALIVE (    "
                         "    TIMESTAMP  INTEGER    "
 						"    );", NULL,NULL,NULL);
-	exitOnError(pSQL3, "create table last_keepalive", sqerr);
+	exitOnError(LOGFD,pSQL3, "create table last_keepalive", sqerr);
 
 	// connection events 
 	sqerr=sqlite3_exec(pSQL3, "CREATE TABLE IF NOT EXISTS EVENTS  (    "
                         "    TIMESTAMP  	INTEGER,    "
                         "    DESCRIPTION	VARCHAR     "
 						"    );", NULL,NULL,NULL);
-	exitOnError(pSQL3, "create table session_events", sqerr);
+	exitOnError(LOGFD,pSQL3, "create table session_events", sqerr);
 
 	// peer table 
 	sqerr=sqlite3_exec(pSQL3, "CREATE TABLE IF NOT EXISTS PEER_INFO (    "
                         "    ADDRESS  	    VARCHAR PRIMARY KEY,    "
                         "    DESCRIPTION	VARCHAR     "
 						"    );", NULL,NULL,NULL);
-	exitOnError(pSQL3, "create table peerinfo", sqerr);
+	exitOnError(LOGFD,pSQL3, "create table peerinfo", sqerr);
 
 
 	// insert peer info 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,
 		"INSERT OR REPLACE INTO PEER_INFO (ADDRESS, DESCRIPTION) VALUES (?,?); ",
 		-1, &pstmt, NULL);
-	exitOnError(pSQL3, "insert/peerinfo/prepare",  sqerr);
+	exitOnError(LOGFD,pSQL3, "insert/peerinfo/prepare",  sqerr);
 
 	sqerr = sqlite3_bind_text(pstmt, 1, 
 								(peer[id].af == 4 ? p_tools_ip4str(id, &peer[id].ip4) : p_tools_ip6str(id, &peer[id].ip6)),
@@ -147,14 +161,15 @@ void p_sqldump_open_file(struct peer_t *peer, int id, struct timeval *ts)
 	sqerr = sqlite3_bind_text(pstmt, 2, "peer inserted", -1, NULL);
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "insert/peerinfo/step",  sqerr);
+	exitOnError(LOGFD,pSQL3, "insert/peerinfo/step",  sqerr);
 
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "inser/peerinfo/finalize",  sqerr);
+	exitOnError(LOGFD,pSQL3, "inser/peerinfo/finalize",  sqerr);
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
 
+	log_info( LOGFD ,  "Created required tables" );
 }
 
 /* log keepalive msg */
@@ -162,11 +177,13 @@ void p_sqldump_add_keepalive(struct peer_t *peer, int id, struct timeval *ts)
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "keepalive/open", sqerr);
+	if (returnOnError(LOGFD,pSQL3, "keepalive/open", sqerr,NULL)) return;
+
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
@@ -174,17 +191,16 @@ void p_sqldump_add_keepalive(struct peer_t *peer, int id, struct timeval *ts)
 	struct dump_msg msg;
 	msg.ts   = ((uint64_t)ts->tv_sec);
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  "UPDATE  LAST_KEEPALIVE SET TIMESTAMP = ?; ", -1, &pstmt, NULL);
-	exitOnError(pSQL3, "update/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "update/prepare",  sqerr,pstmt)) return;
 
 	sqerr = sqlite3_bind_int64(pstmt, 1, msg.ts );
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "update/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "update/step",  sqerr,pstmt)) return;
 
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "update/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "update/finalize",  sqerr, pstmt)) return ;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -195,11 +211,12 @@ void p_sqldump_add_close(struct peer_t *peer, int id, struct timeval *ts)
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "bgpclose/open", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "bgpclose/open", sqerr, pstmt)) return;
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
@@ -207,20 +224,19 @@ void p_sqldump_add_close(struct peer_t *peer, int id, struct timeval *ts)
 	struct dump_msg msg;
 	msg.ts   = ((uint64_t)ts->tv_sec);
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  
 		"INSERT INTO EVENTS (TIMESTAMP, DESCRIPTION) VALUES (?,?); ",
 		-1, &pstmt, NULL);
-	exitOnError(pSQL3, "addclose/insert/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addclose/insert/prepare",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_bind_int64(pstmt,  1, msg.ts );
 	sqerr = sqlite3_bind_text(pstmt, 2, "connection close", -1, NULL );
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "addclose/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addclose/step",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "addclose/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addclose/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -231,11 +247,12 @@ void p_sqldump_add_open(struct peer_t *peer, int id, struct timeval *ts)
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "open/db", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "open/db", sqerr, pstmt)) return;
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
@@ -243,18 +260,17 @@ void p_sqldump_add_open(struct peer_t *peer, int id, struct timeval *ts)
 	struct dump_msg msg;
 	msg.ts   = ((uint64_t)ts->tv_sec);
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  "INSERT INTO EVENTS (TIMESTAMP, DESCRIPTION) VALUES (?,?); ", -1, &pstmt, NULL);
-	exitOnError(pSQL3, "addopen/insert/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addopen/insert/prepare",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_bind_int64(pstmt,  1, msg.ts );
 	sqerr = sqlite3_bind_text(pstmt, 2, "connection open", -1 , NULL );
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "addopen/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addopen/step",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "addopen/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "addopen/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -271,17 +287,17 @@ void p_sqldump_add_withdrawn4(struct peer_t *peer, int id, struct timeval *ts, u
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "withdrawn4/open", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn4/open", sqerr, pstmt)) return;
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  "DELETE FROM PREFIX_PATHS_V4 WHERE PREFIX=?; ", -1, &pstmt, NULL);
-	exitOnError(pSQL3, "withdrawn4/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn4/prepare",  sqerr, pstmt)) return;
 
 	char buf[128];
 	struct in_addr addr;
@@ -289,11 +305,11 @@ void p_sqldump_add_withdrawn4(struct peer_t *peer, int id, struct timeval *ts, u
 	sprintf(buf, "%s/%d", inet_ntoa(addr), mask);
 
 	sqerr = sqlite3_bind_text(pstmt, 1, buf,  -1 , NULL );
-	exitOnError(pSQL3, "withdrawn4/bind",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn4/bind",  sqerr, pstmt)) return;
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "withdrawn4/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn4/step",  sqerr, pstmt)) return;
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "withdrawn4/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn4/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -304,27 +320,27 @@ void p_sqldump_add_withdrawn6(struct peer_t *peer, int id, struct timeval *ts, u
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "withdrawn6/open", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn6/open", sqerr, pstmt))return;
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  "DELETE FROM PREFIX_PATHS_V6 WHERE PREFIX=?; ", -1, &pstmt, NULL);
-	exitOnError(pSQL3, "withdrawn6/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn6/prepare",  sqerr, pstmt)) return;
 
 	char buf[128];
 	inet_ntop(AF_INET6,prefix,buf,INET6_ADDRSTRLEN);
 
 	sqerr = sqlite3_bind_text(pstmt, 1, buf,  -1 , NULL );
-	exitOnError(pSQL3, "withdrawn6/bind",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn6/bind",  sqerr, pstmt)) return;
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "withdrawn6/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn6/step",  sqerr, pstmt)) return;
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "withdrawn6/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "withdrawn6/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -341,12 +357,13 @@ void p_sqldump_add_announce4(struct peer_t *peer, int id, struct timeval *ts,
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	char aspathbuf[1024], prefixbuf[128], nexthopbuf[128];
 	p_sqldump_check_file(peer,id,ts);
 
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "bgpclose/open", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "bgpclose/open", sqerr, pstmt)) return;
 	sqlite3_busy_timeout(pSQL3, 10000);
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
 
@@ -399,11 +416,10 @@ void p_sqldump_add_announce4(struct peer_t *peer, int id, struct timeval *ts,
 
 
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  
 		"INSERT OR REPLACE  INTO PREFIX_PATHS_V4  (PREFIX,ASPATH,COMMUNITYPATH,NEXTHOP,TIMESTAMP ) VALUES (?,?,?, ?,?); ",
 		-1, &pstmt, NULL);
-	exitOnError(pSQL3, "announcev4/insert/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev4/insert/prepare",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_bind_text(pstmt,  1, prefixbuf,    -1, NULL);
 	sqerr = sqlite3_bind_text(pstmt,  2, aspathbuf,    -1, NULL);
@@ -412,9 +428,9 @@ void p_sqldump_add_announce4(struct peer_t *peer, int id, struct timeval *ts,
 	sqerr = sqlite3_bind_int64(pstmt, 5, msg.ts );
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "announcev4/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev4/step",  sqerr, pstmt)) return;
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "announcev4/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev4/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
@@ -431,14 +447,18 @@ void p_sqldump_add_announce6(struct peer_t *peer, int id, struct timeval *ts,
 {
 	int sqerr = SQLITE_OK;
 	sqlite3 * pSQL3=NULL;
+	sqlite3_stmt  * pstmt=NULL;
 
 	char aspathbuf[1024], prefixbuf[128], nexthopbuf[128];
 	p_sqldump_check_file(peer,id,ts);
 	
 	sqerr = sqlite3_open_v2( peer[id].sqldbname, &pSQL3, SQLITE_OPEN_READWRITE, NULL);
-	exitOnError(pSQL3, "announcev6/open", sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev6/open", sqerr,pstmt)) return;
+
 	sqlite3_busy_timeout(pSQL3, 10000);
+
 	sqerr = sqlite3_exec( pSQL3, "BEGIN TRANSACTION", NULL, NULL, NULL );
+	if(returnOnError(LOGFD,pSQL3, "announcev6/begintrans", sqerr,pstmt)) return;
 
 	peer[id].empty = 0;
 		struct dump_msg                     msg;
@@ -481,11 +501,10 @@ void p_sqldump_add_announce6(struct peer_t *peer, int id, struct timeval *ts,
 		inet_ntop(AF_INET6,nexthop,nexthopbuf,INET6_ADDRSTRLEN);
 
 
-	sqlite3_stmt  * pstmt;
 	sqerr = sqlite3_prepare_v2( pSQL3,  
 		"INSERT OR REPLACE  INTO PREFIX_PATHS_V6  (PREFIX,ASPATH,COMMUNITYPATH,NEXTHOP,TIMESTAMP ) VALUES (?,?,?, ?,?); ",
 		-1, &pstmt, NULL);
-	exitOnError(pSQL3, "announcev6/insert/prepare",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev6/insert/prepare",  sqerr, pstmt )) return;
 
 	sqerr = sqlite3_bind_text(pstmt,  1, prefixbuf,    -1, NULL);
 	sqerr = sqlite3_bind_text(pstmt,  2, aspathbuf,    -1, NULL);
@@ -494,9 +513,9 @@ void p_sqldump_add_announce6(struct peer_t *peer, int id, struct timeval *ts,
 	sqerr = sqlite3_bind_int64(pstmt, 5, msg.ts );
 
 	sqerr = sqlite3_step( pstmt);
-	exitOnError(pSQL3, "announcev6/step",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev6/step",  sqerr, pstmt)) return;
 	sqerr = sqlite3_finalize( pstmt);
-	exitOnError(pSQL3, "announcev6/finalize",  sqerr);
+	if(returnOnError(LOGFD,pSQL3, "announcev6/finalize",  sqerr, pstmt)) return;
 
 	sqerr = sqlite3_exec( pSQL3, "COMMIT TRANSACTION", NULL, NULL, NULL );
 	sqlite3_close(pSQL3);
